Remove conflicting dir names
